requirements.txt
hmmlearn
psutil
uv
polars
pyarrow
pandas
google-cloud-storage
google-cloud-bigquery
tqdm
python-dotenv
requests
